uiCommand: move some to ./ui/
