+++ src/code/dialogs/autodraws/homogeneous.js
@@ -8,10 +8,11 @@ import {
   dist2,
   fieldCenter,
 } from "../../geo";
-import { clearAllLinks, getAllPortalsOnScreen } from "../../uiCommands";
+import { clearAllLinks } from "../../ui/operation";
 import wX from "../../wX";
 
 import { displayError, displayWarning } from "../../error";
+import { getAllPortalsOnScreen } from "../../ui/portal";
 
 /**
  * Split a set of portals inside a field into three sets with respect to a portal in the field
