+++ src/code/dialogs/autodraws/onionfield.js
@@ -2,9 +2,10 @@ import { AutoDraw } from "./tools";
 import { WasabeePortal } from "../../model";
 import { getSelectedOperation } from "../../selectedOp";
 import { portalInField } from "../../geo";
-import { clearAllLinks, getAllPortalsOnScreen } from "../../uiCommands";
+import { clearAllLinks } from "../../ui/operation";
 import wX from "../../wX";
 import { displayError } from "../../error";
+import { getAllPortalsOnScreen } from "../../ui/portal";
 
 /**
  * Sort trangle vertices by widest angle
