+++ src/code/init.ts
@@ -18,7 +18,8 @@ import { initFirebase } from "./firebase/init";
 import { onMessage as fbMessageHandler } from "./firebase/event";
 import { postToFirebase } from "./firebase/logger";
 import { initWasabeeD } from "./wd";
-import { listenForPortalDetails, sendLocation } from "./uiCommands";
+import { sendLocation } from "./uiCommands";
+import { listenForPortalDetails } from "./ui/portal";
 import { initSkin, changeSkin } from "./skin";
 import { WPane } from "./leafletClasses";
 import OperationChecklist from "./dialogs/checklist";
