--- src/code/ui/agent.d.ts
@@ -1,8 +1,0 @@
-import { WasabeeAgent } from "../model";
-
-export function formatDisplay(
-  agent: WasabeeAgent,
-  teamID?: number | string
-): HTMLAnchorElement;
-
-export function timeSinceformat(agent: WasabeeAgent): string;
