--- src/code/ui/link.d.ts
@@ -1,13 +1,0 @@
-import type WasabeeOp from "../model/operation";
-import type WasabeeLink from "../model/link";
-
-export function displayFormat(
-  link: WasabeeLink,
-  operation: WasabeeOp,
-  smallScreen?: boolean
-): HTMLDivElement;
-
-export function minLevel(
-  link: WasabeeLink,
-  operation: WasabeeOp
-): HTMLSpanElement;
