--- src/code/ui/portal.d.ts
@@ -1,17 +1,0 @@
-import { WasabeePortal } from "../model";
-import { IITC } from "../../types/iitc";
-
-export function fromIITC(p: IITC.Portal): WasabeePortal;
-
-export function team(portal: any): string;
-
-export function displayName(portal: any): any;
-
-export function displayFormat(
-  portal: any,
-  shortName?: boolean
-): HTMLAnchorElement;
-
-export function get(id: any): WasabeePortal;
-
-export function getSelected(): WasabeePortal;
