+++ src/code/uiCommands.ts
@@ -29,220 +29,6 @@ import { constants } from "./static";
 import type { Wasabee } from "./init";
 import { getMe } from "./model/cache";
 
-export function addPortal(operation: WasabeeOp, portal: WasabeePortal) {
-  if (!portal) {
-    displayError(wX("SELECT PORTAL"));
-    return;
-  }
-  operation.addPortal(portal);
-}
-
-export function swapPortal(operation: WasabeeOp, portal: WasabeePortal) {
-  const selectedPortal = PortalUI.getSelected();
-  if (!selectedPortal) {
-    displayError(wX("SELECT PORTAL"));
-    return;
-  }
-  if (portal.id === selectedPortal.id) {
-    displayError(wX("SELF SWAP"));
-    return;
-  }
-
-  const pr = L.DomUtil.create("div", null);
-  pr.textContent = wX("SWAP PROMPT");
-  pr.appendChild(PortalUI.displayFormat(portal));
-  L.DomUtil.create("span", null, pr).textContent = wX("SWAP WITH");
-  pr.appendChild(PortalUI.displayFormat(selectedPortal));
-  L.DomUtil.create("span", null, pr).textContent = "?";
-  const con = new ConfirmDialog({
-    title: wX("SWAP TITLE"),
-    label: pr,
-    type: "anchor",
-    callback: () => {
-      operation.swapPortal(portal, selectedPortal);
-    },
-  });
-  con.enable();
-}
-
-export function deletePortal(operation: WasabeeOp, portal: WasabeePortal) {
-  const pr = L.DomUtil.create("div", null);
-  pr.textContent = wX("DELETE ANCHOR PROMPT");
-  pr.appendChild(PortalUI.displayFormat(portal));
-  const con = new ConfirmDialog({
-    title: wX("DELETE ANCHOR TITLE"),
-    label: pr,
-    type: "anchor",
-    callback: () => {
-      operation.removeAnchor(portal.id);
-      // window.map.fire("wasabee:crosslinks"); -- only needed if we also reset the cache first
-    },
-  });
-  con.enable();
-}
-
-export function deleteMarker(
-  operation: WasabeeOp,
-  marker: WasabeeMarker,
-  portal: WasabeePortal
-) {
-  const pr = L.DomUtil.create("div", null);
-  pr.textContent = wX("DELETE MARKER PROMPT");
-  pr.appendChild(PortalUI.displayFormat(portal));
-  const con = new ConfirmDialog({
-    title: wX("DELETE MARKER TITLE"),
-    label: pr,
-    type: "marker",
-    callback: () => {
-      operation.removeMarker(marker);
-      window.map.fire("wasabee:crosslinks");
-    },
-  });
-  con.enable();
-}
-
-export function clearAllItems(operation: WasabeeOp) {
-  const con = new ConfirmDialog({
-    title: wX("dialog.clear_all.title", { opName: operation.name }),
-    label: wX("dialog.clear_all.text", { opName: operation.name }),
-    type: "operation",
-    callback: () => {
-      operation.clearAllItems();
-      window.map.fire("wasabee:crosslinks");
-    },
-  });
-  con.enable();
-}
-
-export function clearAllLinks(operation: WasabeeOp) {
-  const con = new ConfirmDialog({
-    title: wX("dialog.clear_links.title", { opName: operation.name }),
-    label: wX("dialog.clear_links.text", { opName: operation.name }),
-    type: "operation",
-    callback: () => {
-      operation.clearAllLinks();
-      window.map.fire("wasabee:crosslinks");
-    },
-  });
-  con.enable();
-}
-
-export function clearAllMarkers(operation: WasabeeOp) {
-  const con = new ConfirmDialog({
-    title: wX("dialog.clear_markers.title", { opName: operation.name }),
-    label: wX("dialog.clear_markers.text", { opName: operation.name }),
-    type: "operation",
-    callback: () => {
-      operation.clearAllMarkers();
-      window.map.fire("wasabee:crosslinks");
-    },
-  });
-  con.enable();
-}
-
-export function listenForAddedPortals(newPortal) {
-  if (!newPortal.portal.options.data.title) return;
-
-  const op = getSelectedOperation();
-  const portal = PortalUI.fromIITC(newPortal.portal);
-  op.updatePortal(portal);
-  WasabeeBlocker.updatePortal(op, portal).then((r) => {
-    if (r) window.map.fire("wasabee:crosslinks:update");
-  });
-}
-
-export function listenForPortalDetails(e) {
-  if (!e.success) return;
-  const portal = new WasabeePortal({
-    id: e.guid,
-    name: e.details.title,
-    lat: (e.details.latE6 / 1e6).toFixed(6),
-    lng: (e.details.lngE6 / 1e6).toFixed(6),
-  });
-  const op = getSelectedOperation();
-  op.updatePortal(portal);
-  WasabeeBlocker.updatePortal(op, portal).then((r) => {
-    if (r) window.map.fire("wasabee:crosslinks:update");
-  });
-}
-
-// This is what should be called to add to the queue
-// can take either an entire array of portal GUID or a single GUID
-// this depends on something listening for the IITC PortalDetailsLoaded hook to process the result
-// see listenForPortalDetails above
-export function getPortalDetails(guid) {
-  if (Array.isArray(guid)) {
-    window.plugin.wasabee.portalDetailQueue =
-      window.plugin.wasabee.portalDetailQueue.concat(guid);
-  } else {
-    window.plugin.wasabee.portalDetailQueue.push(guid);
-  }
-
-  const rate =
-    localStorage[
-      window.plugin.wasabee.static.constants.PORTAL_DETAIL_RATE_KEY
-    ] || 1000;
-
-  // if not already processing the queue, start it
-  if (!window.plugin.wasabee.portalDetailIntervalID) {
-    window.plugin.wasabee.portalDetailIntervalID = window.setInterval(
-      pdqDoNext,
-      rate
-    );
-    console.log(
-      "starting portal details request queue: " +
-        window.plugin.wasabee.portalDetailIntervalID
-    );
-  }
-}
-
-function pdqDoNext() {
-  const p = window.plugin.wasabee.portalDetailQueue.shift();
-
-  // are we done?
-  if (p === undefined) {
-    console.debug(
-      "closing portal details request queue: " +
-        window.plugin.wasabee.portalDetailIntervalID
-    );
-    window.clearInterval(window.plugin.wasabee.portalDetailIntervalID);
-    window.plugin.wasabee.portalDetailIntervalID = null;
-    return;
-  }
-
-  if (p.length != 35) return; // ignore faked ones from DrawTools imports and other garbage
-  // this is the bit everyone is so worried about
-  window.portalDetail.request(p);
-}
-
-// load faked op portals
-export function loadFaked(operation: WasabeeOp, force = false) {
-  const flag =
-    localStorage[window.plugin.wasabee.static.constants.AUTO_LOAD_FAKED] ||
-    false;
-
-  // local storage always returns as string
-  if (flag !== "true" && !force) return;
-
-  const f = [];
-  for (const x of operation.fakedPortals) f.push(x.id);
-  if (f.length > 0) getPortalDetails(f);
-}
-
-// load faked blocker portals
-export async function loadBlockerFaked(operation: WasabeeOp, force = false) {
-  const flag =
-    localStorage[window.plugin.wasabee.static.constants.AUTO_LOAD_FAKED] ||
-    false;
-
-  // local storage always returns as string
-  if (flag !== "true" && !force) return;
-
-  const bp = await WasabeeBlocker.getPortals(operation);
-  const f = bp.filter((p) => p.id === p.name).map((p) => p.id);
-  if (f.length > 0) getPortalDetails(f);
-}
-
 export function sendLocation() {
   if (!WasabeeMe.isLoggedIn()) return;
   const sl =
@@ -265,65 +51,6 @@ export function sendLocation() {
       console.error(err);
     }
   );
-}
-
-export function getAllPortalsOnScreen(operation: WasabeeOp) {
-  const bounds = window.map.getBounds();
-  const x = [];
-  for (const portal in window.portals) {
-    if (bounds.contains(window.portals[portal].getLatLng())) {
-      if (
-        operation.containsMarkerByID(
-          window.portals[portal].options.guid,
-          WasabeeMarker.constants.MARKER_TYPE_EXCLUDE
-        )
-      )
-        continue;
-      const wp = PortalUI.fromIITC(window.portals[portal]);
-      if (wp) x.push(wp);
-    }
-  }
-  return x;
-}
-
-export function getAllPortalsLinked(
-  operation: WasabeeOp,
-  originPortal: WasabeePortal
-) {
-  const x = [];
-  for (const link in window.links) {
-    const p = window.links[link];
-
-    if (
-      operation.containsLinkFromTo(p.options.data.oGuid, p.options.data.dGuid)
-    )
-      continue;
-
-    const linkPortal1 = new WasabeePortal({
-      id: p.options.data.oGuid,
-      lat: (p.options.data.oLatE6 / 1e6).toFixed(6),
-      lng: (p.options.data.oLngE6 / 1e6).toFixed(6),
-      name: p.options.data.oGuid,
-      comment: "in",
-    });
-
-    const linkPortal2 = new WasabeePortal({
-      id: p.options.data.dGuid,
-      lat: (p.options.data.dLatE6 / 1e6).toFixed(6),
-      lng: (p.options.data.dLngE6 / 1e6).toFixed(6),
-      name: p.options.data.dGuid,
-      comment: "out",
-    });
-
-    if (linkPortal1.id === originPortal.id) {
-      x.push(linkPortal2);
-    }
-    if (linkPortal2.id === originPortal.id) {
-      x.push(linkPortal1);
-    }
-  }
-  // console.log(x);
-  return x;
 }
 
 // recursive function to auto-mark blockers
@@ -623,29 +350,4 @@ export function clearAllData() {
     },
   });
   con.enable();
-}
-
-export function setMarkersToZones() {
-  const op = getSelectedOperation();
-
-  op.startBatchMode();
-  for (const m of op.markers) {
-    const ll = op.getPortal(m.portalId).latLng;
-
-    const zone = op.determineZone(ll);
-    op.setZone(m, zone);
-  }
-  op.endBatchMode();
-}
-
-export function setLinksToZones() {
-  const op = getSelectedOperation();
-
-  op.startBatchMode();
-  for (const l of op.links) {
-    const ll = op.getPortal(l.fromPortalId).latLng;
-    const zone = op.determineZone(ll);
-    op.setZone(l, zone);
-  }
-  op.endBatchMode();
 }
